Dayana python lab/6-Learn-How-Code/0-dayana-members/1-mr-rokni/0-toDoList-with-django/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
credentials==1.1
Django==4.0.5
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
Pillow==9.1.1
pytz==2022.7
requests==2.28.1
scraper==0.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
